requirements.txt
# Requirement for python 3.12
# Window 11
# Intel core i5 gen 11
# RAM 16GB

python-dotenv==1.0.1

selenium==4.26.1
fake_useragent==1.5.1
beautifulsoup4==4.12.3
requests==2.32.3

openai==1.55.3
llama-index==0.12.2
sentence-transformers==3.3.1
gpt4all==2.8.2
#langchain
pprintpp==0.4.0
#nomic==3.3.3
#transformers==4.46.3
#torch==2.5.1
einops==0.8.0
memory_profiler 
qdrant-client

# FastAPI
fastapi["standard"]
python-multipart==0.0.20
aiofiles==24.1.0